+++ core/src/textures/bump_texture.h
@@ -15,7 +15,7 @@ public:
 
 	color value(double u, double v, const point3& p) const;
 
-	vector3 perturb_normal(const vector3& normal, float u, float v, const vector3& p) const;
+	vector3 perturb_normal(const vector3& normal, double u, double v, const vector3& p) const;
 	vector3 convert_to_world_space(const vector3& tangentSpaceNormal, const vector3& originalNormal) const;
 	void compute_tangent_space(const vector3& normal, vector3& tangent, vector3& bitangent) const;
